requirements.txt
Pillow
rembg
sqlalchemy
datetime
fastapi
onnxruntime